requirements.txt
paho-mqtt
streamlit
pandas
plotly
